requirements.txt
Flask
beautifulsoup4
requests
python-dotenv
google-generativeai

